+++ app/src/main/java/com/sopt/dive/data/api/AuthService.kt
@@ -3,6 +3,7 @@ package com.sopt.dive.data.api
 import com.sopt.dive.data.dto.RequestLoginDto
 import com.sopt.dive.data.dto.ResponseLoginDto
 import retrofit2.Call
+import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
 
@@ -10,5 +11,5 @@ interface AuthService {
     @POST("api/v1/auth/login")
     fun login(
         @Body request: RequestLoginDto
-    ): Call<ResponseLoginDto>
+    ): Response<ResponseLoginDto>
 }
